+++ lib/td_dq_web/controllers/rule_implementation_controller.ex
@@ -220,7 +220,12 @@ defmodule TdDqWeb.RuleImplementationController do
   end
 
   def show(conn, %{"id" => id}) do
-    rule_implementation = Rules.get_rule_implementation!(id)
+    rule_implementation =
+      id
+      |> Rules.get_rule_implementation!()
+      |> add_rule_results()
+      |> add_last_rule_result()
+
     user = conn.assigns[:current_resource]
 
     with true <- can?(user, show(rule_implementation)) do
@@ -394,5 +399,10 @@ defmodule TdDqWeb.RuleImplementationController do
       :_last_rule_result_,
       Rules.get_last_rule_result(rule_implementation.implementation_key)
     )
+  end
+
+  defp add_rule_results(rule_implementation) do
+    rule_implementation
+    |> Map.put(:all_rule_results, Rules.get_rule_implementation_results(rule_implementation.implementation_key))
   end
 end