+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-1656] Added all rule implementation results to rule implementation show response
+
 ## [3.3.0] 2019-08-05
 
 ### Added